requirements.txt
pandas
sqlalchemy
psycopg2-binary
pydantic
faker
pytest
apache-airflow
